Move rom_map and ptr_table to use usize.
Makes indexing [u8] easier.

+++ src/field/npc.rs
@@ -31,7 +31,7 @@ pub struct Npc {
 
 pub fn ptr_table(rom_data: &[u8]) -> Result<ptr_table::Table, Box<Error>> {
     // TODO bounds check rom data.
-    let addr = rom_map::snes_to_file(rom_map::NPC_POINTERS) as usize;
+    let addr = rom_map::snes_to_file(rom_map::NPC_POINTERS);
     let table = ptr_table::Table::new(
         &rom_data[addr..],
         0x1a0,
